Karabiner complex modification for Shift+Delete
A Karabiner complex modification that turns Shift+Delete into Option+Delete (to be clear, this makes it so that when you push Shift+Delete, it will delete an entire word, but Option+Delete will still do this). I use DeleteCut (by r_plus on BigBoss) on iOS, and have grown used to pushing Shift+Delete to delete an entire word. This lets me do that.